fix(cli): use same process when executing scripts
