+++ packages/sanity/src/cli/threads/registerBrowserEnv.ts
@@ -1,4 +1,4 @@
-import {mockBrowserEnvironment} from '../../util/mockBrowserEnvironment'
+import {mockBrowserEnvironment} from '../util/mockBrowserEnvironment'
 
 mockBrowserEnvironment(
   // eslint-disable-next-line no-process-env
